add import_helper submodule in test.support
